refactor(aina): extract BitmaskGattPttReader base from PrymeBleReader
Moves the vendor-agnostic BLE GATT plumbing — bounded direct-retry
burst then long-lived autoConnect for out-of-range recovery, service
discovery against a list of candidate service UUIDs, broad CCCD
subscription on every notify-capable characteristic in the matched
service, edge-triggered per-button down/up derivation — into a new
[BitmaskGattPttReader] base class configured by a [Config] data
class carrying (log tag, candidate service UUIDs, mask decoder).

PrymeBleReader now consists solely of its Pryme-specific config:
the HM-10 transparent-UART service UUID that production BT-PTT-Z
units actually advertise (0000ffe0-…) with the extracted VX-dex
UUID (00420000-…-b617e493) as a fallback, and a top-level
`decodePrymeMask` function that preserves the current "any non-zero
first byte = PTT held" heuristic. VoicePlant's PrymeBleReader
constructor signature is unchanged so no call-site edits are
needed.

The extraction generalizes the on-the-wire model that PrymeBleReader
was already implementing (GATT connect + notify-subscribe + bitmask
decode) so a second reader for any button family that speaks the
same protocol shape only has to supply its config. That "second
reader" is Zello-convention BLE PTT (see the follow-up commit) —
the split of edge-triggering out of the vendor-specific reader lets
Zello's multi-button decoder produce independent per-button events
without duplicating the diff logic.

Behavior preserved:
  - Same service UUID priority order for Pryme (HM-10 first,
    VX-vendor fallback).
  - Same "any non-zero = PTT" heuristic (verified across 0x01..0xFF
    in the unit tests).
  - Same retry ladder (4 direct attempts with backoff up to 16 s,
    then autoConnect at 5 s).
  - Same CCCD-write path with the Tiramisu / pre-Tiramisu split.

Tests
  BitmaskGattPttReaderTest covers:
    - diffEdges pure function: no-op on identical held sets;
      down-only when the set grows; up-only when it shrinks; both
      when it rotates; and independent per-button transitions when
      the set is a superset / subset.
    - decodePrymeMask: 0x00 → empty; every non-zero byte in
      0x01..0xFF → { PTT }. Locks the "non-zero = PTT" heuristic
      in place so a future refactor can't accidentally narrow it.
